+++ clients/ops/admin-ui/src/features/common/dropdown/MultiSelectDropdown.tsx
@@ -112,9 +112,6 @@ const MultiSelectDropdown: React.FC<MultiSelectDropdwonProps> = ({
             _active={{
               bg: "none",
             }}
-            _focus={{
-              bg: "none",
-            }}
             _hover={{
               bg: "none",
             }}