+++ clients/ops/admin-ui/src/features/common/dropdown/types.ts
@@ -1,0 +1,14 @@
+export type ItemOption = {
+  /**
+   * Specifies the value to be sent
+   */
+  value: string;
+  /**
+   * Specifies that an option should be disabled
+   */
+  isDisabled?: boolean;
+  /**
+   * Specifies a toolTip should be dislayed when the user hovers over an option
+   */
+  toolTip?: string;
+};