+++ clients/ops/admin-ui/src/features/datastore-connections/constants.ts
@@ -63,25 +63,26 @@ export const CONNECTOR_LOGOS_PATH = "images/connector-logos/";
 /**
  * List of connection type image key/value pairs
  */
-export const ConnectionTypeLogoMap = new Map<ConnectionType | SaasType, string>(
-  [
-    [ConnectionType.MARIADB, "mariadb.svg"],
-    [ConnectionType.MONGODB, "mongodb.svg"],
-    [ConnectionType.MSSQL, "sqlserver.svg"],
-    [ConnectionType.MYSQL, "mysql.svg"],
-    [ConnectionType.POSTGRES, "postgres.svg"],
-    [ConnectionType.REDSHIFT, "redshift.svg"],
-    [ConnectionType.SNOWFLAKE, "snowflake.svg"],
-    [SaasType.HUB_SPOT, "hubspot.svg"],
-    [SaasType.MAILCHIMP, "mailchimp.svg"],
-    [SaasType.OUTREACH, "outreach.svg"],
-    [SaasType.SALES_FORCE, "salesforce.svg"],
-    [SaasType.SEGMENT, "segment.svg"],
-    [SaasType.SENTRY, "sentry.svg"],
-    [SaasType.STRIPE, "stripe.svg"],
-    [SaasType.ZENDESK, "zendesk.svg"],
-  ]
-);
+export const CONNECTION_TYPE_LOGO_MAP = new Map<
+  ConnectionType | SaasType,
+  string
+>([
+  [ConnectionType.MARIADB, "mariadb.svg"],
+  [ConnectionType.MONGODB, "mongodb.svg"],
+  [ConnectionType.MSSQL, "sqlserver.svg"],
+  [ConnectionType.MYSQL, "mysql.svg"],
+  [ConnectionType.POSTGRES, "postgres.svg"],
+  [ConnectionType.REDSHIFT, "redshift.svg"],
+  [ConnectionType.SNOWFLAKE, "snowflake.svg"],
+  [SaasType.HUB_SPOT, "hubspot.svg"],
+  [SaasType.MAILCHIMP, "mailchimp.svg"],
+  [SaasType.OUTREACH, "outreach.svg"],
+  [SaasType.SALES_FORCE, "salesforce.svg"],
+  [SaasType.SEGMENT, "segment.svg"],
+  [SaasType.SENTRY, "sentry.svg"],
+  [SaasType.STRIPE, "stripe.svg"],
+  [SaasType.ZENDESK, "zendesk.svg"],
+]);
 
 /**
  * Fallback connector image path if original src path doesn't exist