+++ clients/ops/admin-ui/src/features/datastore-connections/filters/TestingStatusFilter.tsx
@@ -1,4 +1,5 @@
 import SelectDropdown from "common/dropdown/SelectDropdown";
+import { ItemOption } from "common/dropdown/types";
 import { capitalize } from "common/utils";
 import React, { useMemo } from "react";
 import { useDispatch, useSelector } from "react-redux";
@@ -17,11 +18,11 @@ const TestingStatusFilter: React.FC<TestingStatusFilterProps> = ({ width }) => {
   // eslint-disable-next-line @typescript-eslint/naming-convention
   const { test_status } = useSelector(selectDatastoreConnectionFilters);
 
-  const loadList = (): Map<string, string> => {
-    const list = new Map<string, string>();
+  const loadList = (): Map<string, ItemOption> => {
+    const list = new Map<string, ItemOption>();
     const valuesList = Object.values(TestingStatus).sort();
     valuesList.forEach((value) => {
-      list.set(capitalize(value), value);
+      list.set(capitalize(value), { value });
     });
     return list;
   };