+++ CHANGELOG.md
@@ -27,6 +27,8 @@ The types of changes are:
 * Access and erasure support for Adobe Campaign [#905](https://github.com/ethyca/fidesops/pull/905)
 * Added db vs saas to connection type api [#937](https://github.com/ethyca/fidesops/pull/937)
 * Retry a DSR (FE) [#863](https://github.com/ethyca/fidesops/pull/938)
+* Add a Connection - Select a connector to configure (front end) [#760] (https://github.com/ethyca/fidesops/pull/987)
+* Add a Connection - Front End layout structure [#866] (https://github.com/ethyca/fidesops/pull/987)
 
 ### Changed